requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
requests==2.31.0
SQLAlchemy==2.0.46
openpyxl==3.1.5
APScheduler==3.10.4
aiohttp==3.9.5
# psycopg2-binary fails to build on newer Python versions (e.g. 3.14).
# psycopg3 works with SQLAlchemy via the "postgresql+psycopg" dialect.
psycopg[binary]==3.2.13
